Feature 112427 auto deletion  (#146)
* Pooja | Hive-112427 | Auto deletion scenario for manual closure of visit

* Pooja | Hive0112427 | Ensure drafts are not allowed to be created when patient has no active visit, fix for pop up not showing up when user fills a form data after discarding a previous draft

* Pooja | Hive-112427 | Add null check before calling discard Drafts

+++ ui/app/common/services/formDraftService.js
@@ -23,6 +23,9 @@ angular.module('bahmni.common.services')
         };
 
         var discardDraft = function (patientUuid, providerUuid) {
+            if (!patientUuid || !providerUuid) {
+                return;
+            }
             return $http.delete(formDraftUrl, {
                 params: {
                     patientUuid: patientUuid,
